mem_space: Factor out common code
