+++ include/arch/aarch64/mem_space.hpp
@@ -1,4 +1,3 @@
-
 #ifndef LIBARCH_MEM_SPACE_HPP
 #define LIBARCH_MEM_SPACE_HPP
 
@@ -167,60 +166,10 @@ namespace _detail {
 			return t;
 		}
 	};
-
-	struct mem_space {
-		constexpr mem_space()
-		: _base(0) { }
-
-		constexpr mem_space(uintptr_t base)
-		: _base(base) { }
-
-		mem_space(void *base)
-		: _base(reinterpret_cast<uintptr_t>(base)) { }
-
-		mem_space subspace(ptrdiff_t offset) const {
-			return mem_space(reinterpret_cast<void *>(_base + offset));
-		}
-
-		template<typename RT>
-		void store(RT r, typename RT::rep_type value) const {
-			auto p = reinterpret_cast<typename RT::bits_type *>(_base + r.offset());
-			auto v = static_cast<typename RT::bits_type>(value);
-			mem_ops<typename RT::bits_type>::store(p, v);
-		}
-
-		template<typename RT>
-		typename RT::rep_type load(RT r) const {
-			auto p = reinterpret_cast<const typename RT::bits_type *>(_base + r.offset());
-			auto b = mem_ops<typename RT::bits_type>::load(p);
-			return static_cast<typename RT::rep_type>(b);
-		}
-
-		template<typename RT>
-		void store_relaxed(RT r, typename RT::rep_type value) const {
-			auto p = reinterpret_cast<typename RT::bits_type *>(_base + r.offset());
-			auto v = static_cast<typename RT::bits_type>(value);
-			mem_ops<typename RT::bits_type>::store_relaxed(p, v);
-		}
-
-		template<typename RT>
-		typename RT::rep_type load_relaxed(RT r) const {
-			auto p = reinterpret_cast<const typename RT::bits_type *>(_base + r.offset());
-			auto b = mem_ops<typename RT::bits_type>::load_relaxed(p);
-			return static_cast<typename RT::rep_type>(b);
-		}
-
-	private:
-		uintptr_t _base;
-	};
 }
 
 using _detail::mem_ops;
-using _detail::mem_space;
-
-static constexpr mem_space global_mem{};
 
 } // namespace arch
 
 #endif // LIBARCH_MEM_SPACE_HPP
-
