+++ include/arch/riscv64/mem_space.hpp
@@ -89,41 +89,9 @@ namespace _detail {
 		// TODO: Implement:
 		static uint64_t atomic_exchange(uint64_t *p, uint64_t v);
 	};
-
-	struct mem_space {
-		constexpr mem_space()
-		: _base(0) { }
-
-		mem_space(void *base)
-		: _base(reinterpret_cast<uintptr_t>(base)) { }
-
-		mem_space subspace(ptrdiff_t offset) const {
-			return mem_space(reinterpret_cast<void *>(_base + offset));
-		}
-
-		template<typename RT>
-		void store(RT r, typename RT::rep_type value) const {
-			auto p = reinterpret_cast<typename RT::bits_type *>(_base + r.offset());
-			auto v = static_cast<typename RT::bits_type>(value);
-			mem_ops<typename RT::bits_type>::store(p, v);
-		}
-
-		template<typename RT>
-		typename RT::rep_type load(RT r) const {
-			auto p = reinterpret_cast<const typename RT::bits_type *>(_base + r.offset());
-			auto b = mem_ops<typename RT::bits_type>::load(p);
-			return static_cast<typename RT::rep_type>(b);
-		}
-
-	private:
-		uintptr_t _base;
-	};
 }
 
 using _detail::mem_ops;
-using _detail::mem_space;
-
-static constexpr mem_space global_mem{};
 
 } // namespace arch
