+++ include/arch/x86/mem_space.hpp
@@ -1,4 +1,3 @@
-
 #ifndef LIBARCH_MEM_SPACE_HPP
 #define LIBARCH_MEM_SPACE_HPP
 
@@ -92,43 +91,10 @@ namespace _detail {
 			return v;
 		}
 	};
-
-	struct mem_space {
-		constexpr mem_space()
-		: _base(0) { }
-		
-		mem_space(void *base)
-		: _base(reinterpret_cast<uintptr_t>(base)) { }
-
-		mem_space subspace(ptrdiff_t offset) const {
-			return mem_space(reinterpret_cast<void *>(_base + offset));
-		}
-
-		template<typename RT>
-		void store(RT r, typename RT::rep_type value) const {
-			auto p = reinterpret_cast<typename RT::bits_type *>(_base + r.offset());
-			auto v = static_cast<typename RT::bits_type>(value);
-			mem_ops<typename RT::bits_type>::store(p, v);
-		}
-
-		template<typename RT>
-		typename RT::rep_type load(RT r) const {
-			auto p = reinterpret_cast<const typename RT::bits_type *>(_base + r.offset());
-			auto b = mem_ops<typename RT::bits_type>::load(p);
-			return static_cast<typename RT::rep_type>(b);
-		}
-
-	private:
-		uintptr_t _base;
-	};
 }
 
 using _detail::mem_ops;
-using _detail::mem_space;
-
-static constexpr mem_space global_mem{};
 
 } // namespace arch
 
 #endif // LIBARCH_MEM_SPACE_HPP
-
